% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestors.R
\name{ancestors}
\alias{ancestors}
\title{Return the ancestors of an HPO term}
\usage{
ancestors(term)
}
\arguments{
\item{term}{Character vector, a valid HPO ID}
}
\value{
a character vector where each element is a ancestor term
}
\description{
Return the ancestors of an HPO term
}
\examples{
ancestors("HP:0000001")
\dontrun{
data \%>\% mutate(ancestors = ancestors(term))
}
}
