% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{children}
\alias{children}
\title{Return the children of an HPO term}
\usage{
children(term)
}
\arguments{
\item{term}{Character vector, a valid HPO ID}
}
\value{
a character vector where each element is a child term
}
\description{
Return the children of an HPO term
}
\examples{
children("HP:0000001")
\dontrun{
data \%>\% mutate(children = children(term))
}
}
