% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descendants.R
\name{descendants}
\alias{descendants}
\title{Return the descendants of an HPO term}
\usage{
descendants(term)
}
\arguments{
\item{term}{Character vector, a valid HPO ID}
}
\value{
a character vector where each element is a descendant term
}
\description{
Return the descendants of an HPO term
}
\examples{
descendants("HP:0000001")
\dontrun{
data \%>\% mutate(descendants = descendants(term))
}
}
