% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Return the description of an HPO term}
\usage{
describe(term)
}
\arguments{
\item{term}{Character vector, a valid HPO ID}
}
\value{
a character vector of the description (name) of the HPO ID
}
\description{
Return the description of an HPO term
}
\examples{
describe("HP:0000001")
}
