% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents.R
\name{parents}
\alias{parents}
\title{Return the parents of an HPO term}
\usage{
parents(term)
}
\arguments{
\item{term}{Character vector, a valid HPO ID}
}
\value{
a character vector where each element is a parent term
}
\description{
Return the parents of an HPO term
}
\examples{
parents("HP:0000001")
\dontrun{
data \%>\% mutate(parents = parents(term))
}
}
