% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hpo.R
\name{parse_hpo}
\alias{parse_hpo}
\title{Parse an HPO OBO file}
\usage{
parse_hpo(path)
}
\arguments{
\item{path}{Character vector; filepath to your hp.obo file.}
}
\value{
Tibble of HPO terms, term descriptions, and relations
}
\description{
Parse an HPO OBO file
}
\examples{
\dontrun{
hpo <- parse_hpo("hp.obo")
}
}
