% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Exclude descendants of a set of terms (roots) that are not ancestors of
another set of terms. In other words, this takes a set of terms and returns
the subtrees rooted at each node, pruned to nodes that are ancestors of
the second set of terms.}
\usage{
prune(roots, terms)
}
\arguments{
\item{roots}{Character vector, a set of valid HPO IDs; defines descendants}

\item{terms}{Character vector, a set of valid HPO IDs; defines ancestors}
}
\value{
a character vector where each element is a HPO term
}
\description{
Exclude descendants of a set of terms (roots) that are not ancestors of
another set of terms. In other words, this takes a set of terms and returns
the subtrees rooted at each node, pruned to nodes that are ancestors of
the second set of terms.
}
\examples{
prune(roots = c("HP:0001250", "HP:0001943"), terms = "HP:0002173")
\dontrun{
data \%>\% mutate(test = prune(roots, terms = "HP:0002173"))
}
}
