% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snomed.R
\name{snomed}
\alias{snomed}
\title{Maps an HPO term to SNOMED terms}
\usage{
snomed(term)
}
\arguments{
\item{term}{Character vector, a valid HPO ID}
}
\value{
a character vector where each element is a SNOMED term
}
\description{
Maps an HPO term to SNOMED terms
}
\examples{
snomed("HP:0000003")
\dontrun{
data \%>\% mutate(snomed = snomed(term))
}
}
