% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umls.R
\name{umls}
\alias{umls}
\title{Maps an HPO term to UMLS concepts}
\usage{
umls(term)
}
\arguments{
\item{term}{Character vector, a valid HPO ID}
}
\value{
a character vector where each element is a UMLS concept
}
\description{
Maps an HPO term to UMLS concepts
}
\examples{
umls("HP:0000001")
\dontrun{
data \%>\% mutate(umls = umls(term))
}
}
